Create new test pipe instead of adding crap to exported pipe
